+++ PvpAutoLb/Core/Filters/StatusFilter.cs
@@ -13,4 +13,15 @@ internal static class StatusFilter
         }
         return false;
     }
+
+    public static bool IsImmuneToLb(IBattleChara target)
+    {
+        var list = target.StatusList;
+        for (var i = 0; i < list.Length; i++)
+        {
+            if (list[i] is { StatusId: PvpAutoLbConstants.StatusIds.HallowedGround
+                or PvpAutoLbConstants.StatusIds.UndeadRedemption }) return true;
+        }
+        return false;
+    }
 }